+++ UnitTests/Sync/SyncServiceTests.cs
@@ -18,6 +18,15 @@ namespace UnitTests.Sync
 {
 	public class SyncServiceTests
 	{
+		private void ConfigureValidSettings(Settings settings)
+		{
+			settings.Peloton.Email = "someEmail";
+			settings.Peloton.Password = "somePassword";
+
+			settings.Garmin.Upload = true;
+			settings.Garmin.Email = "someEmail";
+			settings.Garmin.Password = "somePassword";
+		}
 
 		[Test]
 		public async Task SyncAsync_When_PelotonStep_Fails_Should_ReturnCorrectResponse()
@@ -25,10 +34,11 @@ namespace UnitTests.Sync
 			// SETUP
 			var mocker = new AutoMocker();
 
-			var config = new Settings();
 			var service = mocker.CreateInstance<SyncService>();
 			var peloton = mocker.GetMock<IPelotonService>();
 			var db = mocker.GetMock<ISyncStatusDb>();
+			var settings = mocker.GetMock<Settings>();
+			ConfigureValidSettings(settings.Object);
 
 			var syncStatus = new SyncServiceStatus();
 			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
@@ -57,6 +67,8 @@ namespace UnitTests.Sync
 			var peloton = mocker.GetMock<IPelotonService>();
 			var db = mocker.GetMock<ISyncStatusDb>();
 			var converter = mocker.GetMock<IConverter>();
+			var settings = mocker.GetMock<Settings>();
+			ConfigureValidSettings(settings.Object);
 
 			var syncStatus = new SyncServiceStatus();
 			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
@@ -84,12 +96,13 @@ namespace UnitTests.Sync
 			// SETUP
 			var mocker = new AutoMocker();
 
-			var config = new Settings();
 			var service = mocker.CreateInstance<SyncService>();
 			var peloton = mocker.GetMock<IPelotonService>();
 			var db = mocker.GetMock<ISyncStatusDb>();
 			var converter = mocker.GetMock<IConverter>();
 			var garmin = mocker.GetMock<IGarminUploader>();
+			var settings = mocker.GetMock<Settings>();
+			ConfigureValidSettings(settings.Object);
 
 			var syncStatus = new SyncServiceStatus();
 			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
@@ -119,13 +132,14 @@ namespace UnitTests.Sync
 			// SETUP
 			var mocker = new AutoMocker();
 
-			var config = new Settings();
 			var service = mocker.CreateInstance<SyncService>();
 			var peloton = mocker.GetMock<IPelotonService>();
 			var db = mocker.GetMock<ISyncStatusDb>();
 			var converter = mocker.GetMock<IConverter>();
 			var garmin = mocker.GetMock<IGarminUploader>();
 			var fileHandler = mocker.GetMock<IFileHandling>();
+			var settings = mocker.GetMock<Settings>();
+			ConfigureValidSettings(settings.Object);
 
 			var syncStatus = new SyncServiceStatus();
 			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
@@ -149,20 +163,20 @@ namespace UnitTests.Sync
 			fileHandler.Verify(x => x.Cleanup(It.IsAny<string>()), Times.Exactly(3));
 		}
 
-		// TODO: migrate to SyncServiceTests
 		[Test]
 		public async Task SyncAsync_Should_ExcludeInProgress()
 		{
 			// SETUP
 			var mocker = new AutoMocker();
 
-			var config = new Settings();
 			var service = mocker.CreateInstance<SyncService>();
 			var peloton = mocker.GetMock<IPelotonService>();
 			var db = mocker.GetMock<ISyncStatusDb>();
 			var converter = mocker.GetMock<IConverter>();
 			var garmin = mocker.GetMock<IGarminUploader>();
 			var fileHandler = mocker.GetMock<IFileHandling>();
+			var settings = mocker.GetMock<Settings>();
+			ConfigureValidSettings(settings.Object);
 
 			var syncStatus = new SyncServiceStatus();
 			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
@@ -189,6 +203,74 @@ namespace UnitTests.Sync
 			peloton.Verify(x => x.GetRecentWorkoutsAsync(0), Times.Once);
 			converter.Verify(x => x.Convert(It.IsAny<P2GWorkout>()), Times.Once);
 			garmin.Verify(x => x.UploadToGarminAsync(), Times.Once);
+			db.Verify(x => x.UpsertSyncStatusAsync(It.IsAny<SyncServiceStatus>()), Times.Once);
+			fileHandler.Verify(x => x.Cleanup(It.IsAny<string>()), Times.Exactly(3));
+		}
+
+		[Test]
+		public async Task SyncAsync_When_PelotonCredsMissing_Fails_Should_ReturnCorrectResponse([Values(null, "")]string email, [Values(null, "")] string password)
+		{
+			// SETUP
+			var mocker = new AutoMocker();
+
+			var service = mocker.CreateInstance<SyncService>();
+			var peloton = mocker.GetMock<IPelotonService>();
+			var db = mocker.GetMock<ISyncStatusDb>();
+			var settings = mocker.GetMock<Settings>();
+			settings.Object.Peloton.Email = email;
+			settings.Object.Peloton.Password = password;
+
+			var syncStatus = new SyncServiceStatus();
+			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
+
+			// ACT
+			var response = await service.SyncAsync(0);
+
+			// ASSERT
+			response.SyncSuccess.Should().BeFalse();
+			response.PelotonDownloadSuccess.Should().BeFalse();
+			response.Errors.Should().NotBeNullOrEmpty();
+
+			peloton.Verify(x => x.GetRecentWorkoutsAsync(0), Times.Never);
+			db.Verify(x => x.UpsertSyncStatusAsync(It.IsAny<SyncServiceStatus>()), Times.Once);
+		}
+
+		[Test]
+		public async Task SyncAsync_When_GarminCredsMissing_Fails_Should_ReturnCorrectResponse([Values(null, "")] string email, [Values(null, "")] string password)
+		{
+			// SETUP
+			var mocker = new AutoMocker();
+
+			var service = mocker.CreateInstance<SyncService>();
+			var peloton = mocker.GetMock<IPelotonService>();
+			var db = mocker.GetMock<ISyncStatusDb>();
+			var converter = mocker.GetMock<IConverter>();
+			var garmin = mocker.GetMock<IGarminUploader>();
+			var fileHandler = mocker.GetMock<IFileHandling>();
+			var settings = mocker.GetMock<Settings>();
+			settings.Object.Peloton.Email = "someEmail";
+			settings.Object.Peloton.Password = "somePassword";
+			settings.Object.Garmin.Upload = true;
+			settings.Object.Garmin.Email = email;
+			settings.Object.Garmin.Password = password;
+
+			var syncStatus = new SyncServiceStatus();
+			db.Setup(x => x.GetSyncStatusAsync()).Returns(Task.FromResult(syncStatus));
+			peloton.Setup(x => x.GetRecentWorkoutsAsync(0)).ReturnsAsync(new List<RecentWorkout>() { new RecentWorkout() { Status = "COMPLETE", Id = "1" } });
+			peloton.Setup(x => x.GetWorkoutDetailsAsync(It.IsAny<ICollection<RecentWorkout>>())).ReturnsAsync(new P2GWorkout[] { new P2GWorkout() });
+
+			// ACT
+			var response = await service.SyncAsync(0);
+
+			// ASSERT
+			response.SyncSuccess.Should().BeFalse();
+			response.PelotonDownloadSuccess.Should().BeTrue();
+			response.UploadToGarminSuccess.Should().BeFalse();
+			response.Errors.Should().NotBeNullOrEmpty();
+
+			peloton.Verify(x => x.GetRecentWorkoutsAsync(0), Times.Once);
+			converter.Verify(x => x.Convert(It.IsAny<P2GWorkout>()), Times.Once);
+			garmin.Verify(x => x.UploadToGarminAsync(), Times.Never);
 			db.Verify(x => x.UpsertSyncStatusAsync(It.IsAny<SyncServiceStatus>()), Times.Once);
 			fileHandler.Verify(x => x.Cleanup(It.IsAny<string>()), Times.Exactly(3));
 		}