+++ ReleaseNotes.md
@@ -8,3 +8,4 @@
 - [#292] WebUI - Fixing `Failed to get settings from db` error message on fresh install
 - [#296] Docs site fixes
 - [#295] WebUI - Fix documentation link on settings page was incorrect
+- [#294] WebUI - Do not try to fetch data from Peloton if Email or Password is missing